fix paging getting more then 10000 items (pagerduty limit)
Maximum upper limit of classic pagination The REST API permits retrieving a maximum of 10000 records via pagination. That is to say, the sum of the offset and limit parameters cannot exceed 10000, or the REST API will respond with status 400. If the total parameter is greater than 10000, then not all records in the set can be retrieved.

Signed-off-by: Markus Blaschke <mblaschke82@gmail.com>